requirements.txt
yt-dlp
ffmpeg-python
mutagen
requests
psutil
tqdm